chromium/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/html/imports/import-nested-dup-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
CONSOLE WARNING: HTML Imports is deprecated and has now been removed as of M80. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Test that a subimport should be loaded even if is linked as a top-level import.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS marks[i] is expected[i]
FAIL marks[i] should be START:nest-dup-parent.html. Was END  :import-nested-dup.html.
FAIL marks[i] should be START:nest-dup.html (of type string). Was undefined (of type undefined).
FAIL marks[i] should be START:nest-dup-child.html (of type string). Was undefined (of type undefined).
FAIL marks[i] should be END  :nest-dup-child.html (of type string). Was undefined (of type undefined).
FAIL marks[i] should be END  :nest-dup.html (of type string). Was undefined (of type undefined).
FAIL marks[i] should be END  :nest-dup-parent.html (of type string). Was undefined (of type undefined).
FAIL marks[i] should be END  :import-nested-dup.html (of type string). Was undefined (of type undefined).
FAIL marks.length should be 8. Was 2.
PASS successfullyParsed is true

TEST COMPLETE
